+++ CHANGELOG.md
@@ -1,3 +1,9 @@
 ## 0.0.1
 
 * TODO: Describe initial release.
+
+## 0.0.2
+
+* Added Example of data_handler plugin
+* bugfix: Fixed issue with data_handler plugin
+* improved widgets for data_handler